requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
asyncpg==0.29.0
python-dotenv==1.0.0
lightgbm==4.3.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
apscheduler==3.10.4
jinja2==3.1.3
playwright==1.41.0
aiolimiter==1.1.0
tqdm==4.66.1
pydantic-settings==2.1.0
python-multipart==0.0.6 